:test_tube: [#5084] Add performance regression test for form variable sync
We noticed 3min wall clock time when updating a form definition that
results in 7000 form variables being updated (local dev environment),
which is completely unacceptable.

Isolating the culprit/hot code path revealed that for 1000 form
variables about 10-12 seconds was being spent, so a conservative time
limit of 5 seconds is set up in tests to catch future performance
regressions. However, realistically this should really take less than
a second, but nobody like flaky tests so we pick something in the
middle to account for different hardware in CI.

Updates with small diffs should even take substantially less time,
as we can ignore work that needs not to be done and avoid creating
instances in the first place.