:bug: [#5087] Make fixed metadata variables available in JSON dump plugin
Because the initial value of fixed metadata variables is set to an empty list and the fixed metadata variables field in the serializer was not set to read only, they would be overwritten when saving the form, resulting in a permanent empty list of fixed variables.

This fix adds an API endpoint to fetch the default fixed variables dynamically in the frontend. This means they will be available directly when setting up the plugin for the first time.