:bug: [#5089] Preventing case changes in 'service fetch config' api calls
Make sure that `body` and `query_params` on service fetch configuration keep their original case type.

In most cases we want to change the text case types of content, because the frontend and backend expect different case notations. For the service fetch `body` and `query_params` the case type schouldn't change, because this results in logic not being performed.

Backport-Of: open-formulieren/open-forms#5089