:bug: [#5088] Ensure collapsed metadata fieldset is not hidden
When a lot of variables are selected, the metadata fieldset gets pushed down in the modal. If it is collapsed, this causes it to disappear and not be accessible at all. The CSS styling for this collapsed class contains an `overflow: hidden` property that causes this.

This change adds a custom scss class openforms-fieldset, which overrides it to `overflow: visible`