:lipstick: [#4990] Change when plugin fieldset is shown in registration variables table
Previously, the plugin fieldset was only shown when multiple configured backends introduced registration variables. This led to confusion when there were multiple backends configured, but only ONE of them introduced registration variables. In this case, the registration variables would not be grouped by their plugin name, and it was not clear to which backend they belonged.

Now, the fieldset will only be hidden if there is only ONE configured backend, OR if all the configured backends are of the same type (meaning they introduce the same registration variables).