:test_tube: [#5104] Added tests for radio component default value
The test covers the entire bug reproduce flow. Starting by creating a form with a radio component and saving it. This should (also with the bug still present) set the defaultValue to empty string in the database. When editing the form and opening the radio component in the json view, the defaultValue will show as `null`. Because playwright cannot target the json view correctly (certain dom elements aren't targetable), we just save the component (setting the defaultValue to `null`) and save the form. With the bug present, the database should now have the defaultValue as `null`.

This bug is a combination of problems that result in the defaultValue being wrongly set by formio. The problem comes from our radio component, which defines a defaultValue in the schema. This schema is not set on the formio component, unless we specify it with `get defaultSchema`.

When calculating the defaultValue, formio starts with the defaultValue on the schema, which isn't defined in the regular formio radio component schema. If the defaultValue defined on the component (via the admin) is truety, this will be used as 'real' the defaultValue. Because the defaultValue of the component was an empty string, which is falsey, and the defaultValue wasn't set by the component schema, the defaultValue was set to null. If you would edit the component and save it, then the defaultValue `null` would be saved in the database.