:bug: [#5104] Setting textfield and radio defaultSchema and defaultValue properly
The previous solution for a similar bug for the textfield component was a bandage fix (#4659).

The real problem is that formio doesn't normally define defaultValue's. This is something that we have to define and enforce ourselfs. By setting our custom schema (which is the formio schema + some custom schema definition) we can ensure that a proper defaultValue is defined.

This also makes the components more predictable. If for some reason you set the defaultValue to `null`, then this will be respected. With the previous solution, a empty string was forced as the minimum defaultValue (if the defaultValue was falsey, it would be replaced by an empty string)