:children_crossing: Added fix script
The PR for #4810 fixed an issue with the soft-required component being set to validate `required: true` (which locks a form, as the user cannot provide a value for the component).
As part of the patch release, we should provide users/admins/tech people an option that fixes the problem for all affected forms (besides "fix the JSON of all affected components manually").

This script sets the validation rule of `required` for all soft-required components to `false`. Fixing all form definitions affected by this bug