:test_tube: [open-formulieren/security-issues#36] Added api tests for sanitizing the form definition components content
Testing the create and update proces of form definitions

:safety_vest: [open-formulieren/security-issues#36] Backend html content sanitizing for form definition components

When creating or updating a form definition, the components are sanitized on HTML content. Currently only the component tooltip, description and label will be sanitized.

The sanitizing is done using bleach, and uses a list of allowed tags and attributes. Any html that isn't defined in these lists, will be removed. The content of such a disallowed html component will remain

:safety_vest: [open-formulieren/security-issues#36] Frontend HTML sanitizing component data

Added a simplistic sanitize function that sanitizes component data when saving the component. This function removes most HTML content from the label, tooltip, description and placeholder component data. Only basic semantic tags (`a`, `b`, `em`, `i`, `string`, `u`) are allowed. Also the allowed attributes has been limited to `href` on `a` tags.

Backport-of: #5121