:safety_vest: [open-formulieren/security-issues#36] Backend html content sanitizing for form definition components
When creating or updating a form definition, the components are sanitized on HTML content. Currently only the component tooltip, description and label will be sanitized.

The sanitizing is done using bleach, and uses a list of allowed tags and attributes. Any html that isn't defined in these lists, will be removed. The content of such a disallowed html component will remain