[#5124] Added check for hidden/disabled component in ValidatePrefillData
When  a component (or its parent is disabled and hidden, the validation in the backend is
triggered since there is no data for the specific key. This means, for our validator,
that the data has been changed and raises an error.

* Added test for prefill data persistence

This ensures that if a field is prefilled but not hidden and the data is omitted from
the submission data, the initial prefill data is persisted in the database.
So, if a user clears a prefilled field that is read-only we make sure
that this will not affect the real value of the variable.

Backport-of: #5135