:white_check_mark: [#5047] Added pdf generate test for components without content
When a fillable component (i.e. non layout, things that aren't repeatingGroups or fieldsets) is left empty, a "No information provided" will be shown as the provided data for that field.
This shows that this field was left empty deliberately (and not some pdf generate bug/mistake). Definitely for screen-reader users this will give extra context about the submission