:construction: [#5139] Remove DataContainer usage from iter_evaluate_rules
It's a quick test to see what happens if we were to remove the DataContainer from iter_evaluate_rules, and replace it with a FormioData instance.

Also added support for FormioData in recursive_apply to fix several failing tests.

test_two_actions_on_the_same_variable illustrates what happens when two actions are performed on the same variable.