:children_crossing: Declutter variables table a bit more
The component variables are now grouped by the step they occur in, and
the column with the form definition name is removed. Variables are
sorted alphabetically, which should give a stable ordering of variables
inside of a step (or other tab). The steps are of course rendered in
order of definition.

The key of a component is now properly marked up as code, as it is a
system identifier and can help with quickly visual scanning which key
is required in logic rules.