[#5124] Added test for prefill data persistence
This ensures that if a field is prefilled but not hidden and the data is omitted from
the submission data, the initial prefill data is persisted in the database.
So, if a user clears a prefilled field that is read-only we make sure
that this will not affect the real value of the variable.