:bug: [#5142] Ensure variable bulk update endpoint returns list of variables
This bug was introduced with the fix for GH-5058 (commit d86a1bbd), where FormVariableListSerializer no longer processes component variables. This means that after saving the serializer, `serializer.data` only contains user-defined variables,and we have to ensure it returns all form variables with another serializer.

The expected behaviour is that the "resulting" set of form variables (component + user defined) is returned. The endpoint (via PUT) ensures that the stale component & user defined variables are cleaned up, which effectively means we discard the body data w/r to the component variables. Only user defined variable mutations actually have an effect. It's not pretty, but returning all available form variables in a separate serializer instance should do the job. This will also mimic the behaviour of the `FormViewSet.variables_list` api endpoint.

* Update failing test

Because we now return the list of all available form variables in `FormViewSet.variables_bulk_update`, we need to ensure the form does not contain other variables that are initialized elsewhere. Removing the generate_minimal_setup factory trait does the trick.

Backport-of: #5154