:bug: [#5142] Ensure variable bulk update endpoint returns list of variables
This bug was introduced with the fix for GH-5058 (commit d86a1bbd), where FormVariableListSerializer no longer processes component variables. This means that after saving the serializer, `serializer.data` only contains user-defined variables, and we have to manually create a list of all submitted variables to return in the response.