:construction: [#4951] Setup custom WMTS map renderer
A WIP setup for a custom WMTS map renderer, using geojson (converted to the RD (Rijksdriekhoek) coordinates system) as data input. The function creates a map img, which is then base64 encoded and placed in the html content of the summary pdf.

This setup is a WIP. Meaning, it works but is definitly not nice! All its features are spead-out across two files, and multiple small functions. Key is to simplify the setup, and reduce the amount of "noice" (move all the calculations and moving parts to a central "generate map" place).