:poop: [#5162] Don't error hard on API standard violations
Before this patch, the behaviour was to outright reject non-standard
compliant responses, since the API specifications for the ZGW API's
(managed by VNG) are strictly versioned according to semver. Certain
features are available only starting with certain versions, and we
adapt our client behaviour to be as efficient as possible based on this
API version. It is nonsense for an API server to reply with a version
number that is not published by VNG on the documentation site
https://vng-realisatie.github.io/gemma-zaken/standaard/.

That said, Dimpact is in a pickle because a vendor is not behaving
according to the standard, triggering this hard rejection in Open Forms,
and said vendor is not able to patch their own systems in time for a
particular release time line. We agreed to coalesce the version to the
lowest possible supported so that the application can actually be used
still, however this triggers all the worst performance and UX paths.

This patch is *only* applied to the stable/2.8.x branch since we really
want to bring the ecosystem to standard-compliant implementations.