:card_file_box: [#5158] Remove obsolete model
This model was deprecated in the 2.7.x release cycle and became
obsolete in 2.8.x. However, the ZGW API group cannot be deleted
in the admin because of the model due do how Django's model
deletion collector works. Simply not-blocking the instantiating
does not resolve the issue, since the models.PROTECT on_delete
argument still prevents the admin interface from deleting the
API group and we don't expose the ZGWConfig model any longer
(and deleting it doesn't resolve the issue since it will be
recreated by django-solo anyway).

So, the proper solution is to remove the model and deal with
the migration and the forward-porting required, otherwise
we end up in migration mess anyway.

Forwardport-of: #5165