:green_heart: Fix test flakiness
Avoids a race condition where the DB transaction was not saved yet by
the API call made through the Javascript before asserting the DB state.

The page.wait_for_url was probably a no-op because the page URL is
already satisfied when editing a form and reloading that same page. This
is mitigated by hitting the save that redirects to the form list page,
which has a different URL and provides stronger guarantees that the
changes have been persisted in the database.