:arrow_up: Bump kombu to 5.4+
Kombu is the underlying transport library used by Celery and v5.4.0
should fix the redis stability errors where workers cannot reconnect
after connection loss to redis. It should work with Celery 5.4 and we
should not need to wait for Celery 5.5 to get this fix.

This problem and fix are tough to validate so we'll just have to try
it out.