:test_tube: [#5151] Added test for hidden map component in logic-check call
The logic-check should not return an empty value for map components. The empty value for a map component should be None/null/undefined.

If the logic-check returns any empty value for a map component, this will be merged into the SDK form data. This then would cause validation errors, seeing as a map component should be an object with `type` and `coordinates` properties (with valid values).

Currently, with the map component data type that wasn't correctly changed with the map component re-write, the empty value returned by logic-check is an empty array. Which causes the "Ongeldige data. Verwacht een dictionary, kreeg een list." validation error. When we correct the data type of the map component to object, we would receive a "missing property 'type'" validation error.
The only valid empty value for a map component is *no* empty value.