:recycle: [#5139] Refactor SubmissionValueVariablesState.set_values
Now accepts a FormioData instance, as this is the default data structure during logic evaluation. This is the beginning of changing
DataMapping with FormioData in all function signatures. For now, the FormioData instances are created in form_logic, but in the future it
might be worth considering placing it one level higher.