:bug: [#5151] Map component empty value
The map component should not receive an empty value, if it's missing from the dataset. This is because an empty map component doesn't pass the serializer validations.

A valid map component should be an object with the properties `type` and `coordinates`. If the map component doesn't have these properties, it should fail the validation, as the submitted data cannot be parsed/used correctly.

Without this change, the check-logic would return some empty value for the map component which will then be merged into the SDK form data. If we then submit the form, the validation would break on the map component, because the required properties aren't provided.