:bug: [#5155] Add initial data reference to `onLanguageChange` callback
Previously, the initial data reference would be lost when switching language during filling out a form. This is because, as of GH-4600, we reload the entire page on a language switch to update all links.

Now, the initial data reference is passed to the `onLanguageChange` callback, so we can add it to the link when reloading the form.