:recycle: [#5179] Remove glom from FormioData
Previous fixes (GH-4768 and GH-5118) have shown that the usage of glom in FormioData - especially for large forms - can cause significant performance issues. This is because glom does a lot more than just enabling nested-data access. This patch removes it favour of a custom implementation, as the only thing we need FormioData to do, is this nested-data access.

Note that the fixes applied in GH-4768 and GH-5118 will remain, because they significantly speed up (up to twice as fast) non-nested key access. They maintenance overhead of this extra check is minimal and properly contained in the data structure.