:label: [#5188] Add type annotations so I can reason about the data
Before 3.0.x, we apparently stored a list of attributes to look up (
so a list of strings) for prefill, which we could directly compare
against the prefill config for each component.

The type annotations have been set up for the legacy case (which is
currently broken). What is now really being passed, is a list of dicts
which is wrong.