:vhs: Convert mock-based flaky test to VCR
With the pending zgw-consumers removals and test flakiness, now is as
good a time as any to convert legacy mock-based tests to VCR which
is more robust and more representative. The conversion already
highlighted certain pre-requisites not being met properly, such
as having a BSN for the case initiator or properly versioned
document types etc.