:bento: [#5176] Re-record Referentielijsten API VCR cassettes
The referentielijsten API sends the data (by default) taking into
account the ordering (order_by("-pk")), so according to our fixtures the
results have to be reversed, just like Robin did in the initial commit.

I have just added the new tests that are now in master for the email
digest.