:technologist: Allow skipping e2e tests in PRs/commits
Use the `[skip: e2e]` fragment in the PR body or, when pushing a
commit directly, in the commit message.

Note that you can't include it in a commit message for a PR, as this
is not exposed in the PR context easily. So it *must* go in the body
of the PR description.