:bug: [#4917] Avoid escaping valid design token value characters
The double quote has meaning when specified as a CSS value, don't
escape that. Instead, we now only escape less than and greater than,
as those could affect the CSS parsing and terminate the style tag
before we intend it to.

This is not 100% yet - these less/greater than characters *could* be
valid if they're part of a string to be displayed, but let's wait for
someone to complain about that before we put too much effort in.