:lipstick: [#4917] Fixed the markup and styling for cosign submission lookup
This one was a bit more tricky since we want to have a single container
for the buttons to manage the proper alignment and icons via design
tokens.

Because the submit form and the logout form go to different endpoints,
they need to be their own separate form elements. We can actually put
the submit buttons *outside* of the form elements and still relate them
together, by using the HTML id attribute on each form, and linking the
button via the 'form' attribute to each.

Additionally, there was some invalid markup where block level elements
were present in paragraphs and some p tag was closed with a div tag,
that has been fixed too now.