:bug: [#5181] Fix default values of components
Just like the fixes #4738 and #5123, the default value of these components (checkbox, edit grid, file, and signature) in the form builder was set to `null`. Because this does not match the default values set by the backend - specifically when the component is hidden and 'clearOnHide' is set (see `get_component_empty_value`) - it causes premature validation to be performed.