:vhs: Convert some tests to VCR.py
This migrates the flaky zaakeigenschappen test to use VCR.py and talk to
a real Open Zaak instance instead, for an accurate simulation of actual
behaviour. The mocks are very flawed and converting this revealed a number
of issues (like #5239 and the failing NNP initiator role case).

Some tests are also slimmed down to only assert on the actual different
thing in each case, which enables better focus and avoid assertion
fatigue.

For the expected failure:

We don't actually support this case becausea KVK number is 8 digits,
while the ZGW APIs expect an RSIN which is 9 digits. Open Zaak
(according to the standard) properly rejects this, and leads to a
registration failure. This means that our code (since 2,5 years)
is sort-of broken or at least not usable unless someone jumps through
hoops to get an RSIN in there somehow via the form fields.

These tests were originally already unreliable - they seem to have been
duplicating each other and no assertions were made about the rolType
in the first place, making it unclear if vestiging or niet_natuurlijk_persoon
was supposed to be created. Though, since we don't have a vestigings-
nummer, we can't do more than try to create a NNP role. This *will*
break in production.

This looks to be unresolved in the standard still, see:

* VNG-Realisatie/gemma-zaken#656
* VNG-Realisatie/gemma-zaken#1834
* VNG-Realisatie/gemma-zaken#1885