:fire: [#5221] Remove outdated argument from check_submission_logic
This argument was introduced as a fix for GH-1083, but currently does not serve any purpose. It was added to the context of
``SubmissionStateLogicSerializer`` in ``SubmissionStepViewSet.logic_check``, but all unsaved data is already saved to the state
in ``evaluate_form_logic``. The other two occurrences did not add anything either, as they were both fetching the data from
``Submission.data``, which in turn fetches it from the state. This means data fetched from the state is saved to the state again.

Note that the test that was added with this fix (CheckLogicEndpointTests.test_update_not_applicable_steps) still passes