:fire: [#5221] Remove setting data to step
With the changes from GH-5139, all the data is saved by resolving the submission
value variables state inside ``evaluate_form_logic``. This means setting the data
to the step beforehand is not necessary anymore. Also, including the submission
data in the data sent to ``evaluate_form_logic``, is redundant.  This is because
as ``Submission.data`` just fetches the data from the state, which already happens
inside ``evaluate_form_logic``.