:bug: [#5263] Ensure data in JSON registration is only encoded once
We encode the data to JSON manually using the ``DjangoJSONEncoder`` class to be able to handle date objects, for example.
The data was added to the ``json`` argument of the post request, which is a convenience argument that also encodes it,
meaning it was encoded twice.

Now it is added to the ``data`` argument of the request, and the correct corresponding header is included too.

Note that the VCR cassettes do not need to be re-recorded, as the double decoding was performed in the corresponding flask
app as well.