:sparkles: [#5140] Add form authentication backend configuration
Due to the new upcoming authentication plugins (Yivi and eIDAS) that require form specific configuration, it's best to create a strong setup. Previously we had authentication options and config directly saved on the form object. To make things a bit more structured, we decided to adept a more object-oriented way of working. With FormAuthenticationBackend, similar to FormRegistrationBackend, we can define the auth plugin configuration is a more structured way.