:whale: [#5296] Add JWT_LEEWAY environment variable to Open Zaak docker container
From https://open-zaak.readthedocs.io/en/stable/installation/reference/time.html#jwt-validation:
"Clock drift can cause a JWT that’s valid on the client to be considered not valid (yet, or anymore) on the server, leading to unexpected HTTP 403 errors.

In Open Zaak, you can configure JWT_LEEWAY which allows you to specify the leeway in seconds. This value should be greater than the difference between server and client clock, but be careful to not make the value too large because that compromises security (i.e. - a token is valid longer than intended)."

We add this env variable and set it to a large number to ensure no unexpected errors are raised when running tests.