:sparkles: [#5132] Extend Yivi plugin with DigiD and eHerkenning fields and add Yivi configuration to admin
Add DigiD and eHerkenning config fields to the Yivi configuration, seeing as Yivi can be used to fetch bsn OR kvk. The fields are required, to make sure Yivi can be used in all manner of ways.

Also added configuration for the Yivi plugin. This includes defining available scopes, which will be using for authenitcation. Which scopes should be used, will be defined/configured on form level.