:recycle: [#4933] Bake explicit cosign versioning into the JSONField
It's hard to correctly interpret the data from cosigning when
supporting multiple versions and applying the correct serializer
validation.

The validation serializer is hooked up to be self-documenting,
but this requires the serializer validation to be called too
in production and tests to make sure that it still accurately
documents the shape of data (this was the problem why it
was outdated).

To make the serializer accept multiple versions of the cosign
component/plugin, the easiest path is a polymorphic serializer
that validates the shape based on the version of data it's
receiving, but this requires an explicit discriminator field.
Hence, we add a version field and can add a data migration to
update existing data (because we know roughly which shape
matches which version).