:bug: [#5300] Fix nested data not being saved to the state
Some context:
As of writing this, the ``SubmissionValueVariablesManager.bulk_create_or_update_from_data`` method loops over all
variables and updates them if their key is in the provided data. If a key is not present, the initial value for that
variable is used.

The problem:
With the changes from #5221, ``SubmissionValueVariablesManager.bulk_create_or_update_from_data`` was also affected.
The only direct usage of this in the code is in the ``SubmissionStep.data`` setter. The root cause of this bug is
that when submitted data is passed to the ``SubmissionStepSerializer``, and it is saved, it implicitly sets this
data to ``SubmissionStep.data``, so ``.bulk_create_or_update_from_data`` will be executed. In the PUT call to the
submission step, the data passed to the serializer is directly fetched from the request, so not converted to a
``FormioData`` instance. Therefore, when trying to get a value from the request data for a key with a dot, it will
raise a key error. This means the value of the submission variable is set to the initial value of the form variable,
and no entry is created in the database.

Solution:
We fix this by making ``SubmissionStepSerializer.data`` a ``FormioDataField``, which returns a ``FormioData`` instance
after validation.