:sparkles: [#4923] Update formio configuration before generating form schema
This ensures radio, select, and selectboxes with another variable or reference
lists as a data source will have their options updated.

Note that we no longer perform a check on the data source in the these components
when generating the schema, but rather if the values are available. This is because
these values _will_ be available if the configuration was updated before generating
the schema.