:white_check_mark: [#5285] Remove logging assertions in tests
Discussed with Steven and Sidney - we concluded that the added value of
asserting on certain logging calls in tests is not worth the effort
of converting this to structlog. The reasoning is that if you remember
to test the logging call, you definitely remembered to add the logging
call itself, and if you forget to add the log statements, you'll
probably also forget to test for them.

This does not take away from audit log checks, but in those situations
it's far easier to assert on certain DB records being created and *how*
that happens isn't particularly interesting (be it directly, be it
via the logging framework).