:wrench: [#5285] Format uwsgi server logs as JSON
The bulk of the work can be done in uwsgi config file, in particular
the server logs themselves (startup messages etc) can be passed to the
json-encoder and those will be properly json formatted *and* the
variables will be JSON escaped.

The log lines coming from the application are routed based on their
first character - if it's a curly brace, it's assumed that it's
already (valid) JSON and they are emitted as-is.

Then, uwsgi itself also emits request logs, which is the nasty part.
log-format is the quickest way to get something like JSON outputted,
but it can trip on malformed input (if it has double quotes, or a
comma for status code etc.). This should typically not happen, and
even if it does, we can live with it. The alternative is writing a
C-extension which has been done before but eh... we considered it
not worth the effort.

One particularly tricky bit however is getting the timestamp emitted
in the proper ISO8601 format. You *can* enable this through the
log-format-strftime and log-date configuration options, but this
leads to every server log message (the message!) being prepended
with this timestamp, while it's already emitted separately under
the timestamp key. This is really ugly, and there doesn't seem to
be a way to decouple those.

However, uwsgi injects itself in the python process and you can import
it, to then set log variables in your python code, which is exactly
what's done with the WSGI middleware. The added benefit is that we do
get microsecond precision again, which helps better in correlating logs
with the structlog middleware logs. And, for the future, this will also
allow us to inject/extract the tracing logs/headers to include in the
uwsgi request logs.

Finally, an extra 'source' key is added to the application logs so that
we can easily see whether a log record is coming from the applicaton
itself or the application server, giving a bit more consistent logging
format across the board.