:recycle: [#4923] Refactor post-processing of the generic JSON plugin
Because we now update the formio configuration inside ``generate_json_schema``, there is no
need to do this here as well. Also, processing of radio and select components is no longer
necessary because of this. If the data source is another variable or reference lists, the
options will already be added on a component plugin level (see ``Radio.as_json_schema``, ```Select.as_json_schema``, and ``SelectBoxes.as_json_schema``)

Also removed the redundant creation of a user-defined variable in a test for reference lists