:sparkles: [#5132] Add Yivi options serializer
Adding a Serializer for the Yivi plugin options. This is a first setup, containing the basics for BSN authentication and addition scopes. KVK and pseudo are also allowed as authentication options, but they don't yet have detailed serializers

The choices for additional_scopes is defined dynamically from the Yivi addictional scopes configuration.