:sparkles: [#5132] Extend support of JsonSchemaSerializerMixin to PolymorphicSerializer classes
Add polymorphic shapes, defined in/with drf_polymorphic PolymorphicSerializer, to the JsonSchemaSerializerMixin generate json schemas.

The typical polymorphism of JSON schema is defined with "anyOf", which tells that an object *can* contain certain properties. But "anyOf" tells nothing about when a property/shape should be part of the whole.

To add some relationship between polymorphic discriminator and object shape, a `discriminator` property will be included in the JSON schema. Note: `discriminator` isn't a supported property of JSON schema. The `discriminator` property contains a reference to the "discriminator" field and a mapping of which discriminator value corresponds to which partial schema.