:label: [#5254] Satisfy the type checker
The type checker was complaining because it was essentially getting
hung up on the nominal types defined in the StUF-BG apps and Haal
Centraal variants, which have the same name but a slightly different
shape and they are separate classes.

Instead, the type checker is now nudged towards structural typing via
a protocol that only requires the 'model_dump' method to exists, as
is provided by pydantic models. Both of the nomimal types are
pydantic models and satisfy the protocol, which guarantees that the
resulting output can be JSON-serialized and saved in our JSONB columns.

This also results in quite a bit of a rewrite - the actual first stage
towards serialization is now handled in the plugin itself rather than
in the handler, which obsoletes the 'convert_to_json_serializable'
utility entirely. It was called in many different branches, but those
are also no longer necessary as we can much more easily shortcut.

The repeated pattern of using the options parameter 'type' key as
result dict key was invariant between the different decision branches
within a single handler, *and* even identical to all the handlers, so
it makes more sense to do this operation on the _result_ of the handler
instead of making the handler do this.

This, in turn, makes it easier to simplify reason in the
handler-specific data types and makes things quite a bit more straight
forward.