:lipstick: [#5132] Authentication options forms using Formik
The authentication options have been moved from inline to modal forms. This creates a contained environment for auth plugin configuration. This also reduced the "noice" of the formbuild a bit.

This allows us to use Formik for the auth options forms. Making it easier to support/maintain/develop the authentication configuration.