:green_heart: Ignore about:* urls when capturing requests in e2e tests
These URLs apparently get triggered because of the WYSWIWYG editors
that use a srcdoc attribute in an iframe. These must have a URL,
and internally, browsers assign this the URL 'about:srcdoc' per spec.

These URLs also appear in the network calls of the browser and thus
are picked up by Playwright. We are then in turn trying to resolve that
URL to one of our own views, which fails of course.

These tests should no longer log that they're flaky, but let's observe
for a while.