:bug: Fix hidden selectboxes component returning empty object
If a selectboxes component is hidden with clearOnHide enabled, and uses another variable or
reference lists as a data source, the `defaultValue` is an empty object. If the formio
configuration was updated dynamically, we can generate a correct empty value from the
`values`, e.g. `{"a": False, "b": False, "c": False}`.