:green_heart: Fix tests
The O365 mocking doesn't play nice with static type checking,
so we only assert when the type checker is running and not at
runtime, otherwise tests fail.

The Objects API tests have now been annotated with the proper
plugin options type and the missing keys that are provided by the
options serializer are added to resolve the KeyError because of
invalid options shape being passed.